+++ test/index.js
@@ -411,19 +411,38 @@ test('requesting a page of features that is deflate encoded', function (t) {
 })
 
 test('requesting a page of features and getting an html response', function (t) {
-  var page = new Buffer('</html></html>')
+  var page = Buffer.from('</html></html>')
   var fixture = nock('http://servicesqa.arcgis.com')
 
-  fixture.get('/97KLIFOSt5CxbiRI/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision=')
+  fixture.get('/97KLIFOSt5CxbiRJ/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision=')
   .times(4)
   .reply(200, function () { return page })
 
-  var service = new FeatureService('http://servicesqa.arcgis.com/97KLIFOSt5CxbiRI/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0', {backoff: 1})
-  var task = {req: 'http://servicesqa.arcgis.com/97KLIFOSt5CxbiRI/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision='}
+  var service = new FeatureService('http://servicesqa.arcgis.com/97KLIFOSt5CxbiRJ/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0', {backoff: 1})
+  var task = {req: 'http://servicesqa.arcgis.com/97KLIFOSt5CxbiRJ/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision='}
 
   service._requestFeatures(task, function (err, json) {
     t.ok(err)
     t.equal(err.message, 'Paging aborted: Received HTML or plain text when expecting JSON')
+    t.end()
+  })
+})
+
+test('requesting a page of features and getting an html response tacked on', function (t) {
+  var features = require('./fixtures/features.json')
+  var page = Buffer.from(JSON.stringify(features) + "\n<!DOCTYPE html>")
+  var fixture = nock('http://servicesqa.arcgis.com')
+
+  fixture.get('/97KLIFOSt5CxbiRK/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision=')
+    .times(4)
+    .reply(200, function () { return page })
+
+  var service = new FeatureService('http://servicesqa.arcgis.com/97KLIFOSt5CxbiRK/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0', { backoff: 1 })
+  var task = { req: 'http://servicesqa.arcgis.com/97KLIFOSt5CxbiRK/arcgis/rest/services/QA_data_simple_point_5000/FeatureServer/0/query?outSR=4326&f=json&outFields=*&where=1=1&resultOffset=1000&resultRecordCount=1000&geometry=&returnGeometry=true&returnZ=true&geometryPrecision=' }
+
+  service._requestFeatures(task, function (err, json) {
+    t.notOk(err)
+    t.equal(json.features.length, 1)
     t.end()
   })
 })